+++ src/provider/Hover/tools/HoverDirectives.ts
@@ -1,4 +1,5 @@
 import type * as vscode from 'vscode';
+import { CAhkInclude } from '../../../AhkSymbol/CAhkInclude';
 import { CAhkDirectives } from '../../../AhkSymbol/CAhkLine';
 import type { TAhkSymbolList, TAstRoot } from '../../../AhkSymbol/TAhkSymbolIn';
 import { DirectivesMDMap } from '../../../tools/Built-in/Directives';
@@ -6,10 +7,10 @@ import { DirectivesMDMap } from '../../../tools/Built-in/Directives';
 function findDirectivesWithPos(
     AST: Readonly<TAhkSymbolList>,
     position: vscode.Position,
-): CAhkDirectives | undefined {
+): CAhkDirectives | CAhkInclude | undefined {
     for (const ah of AST) {
         if (ah.range.contains(position)) {
-            return ah instanceof CAhkDirectives
+            return ah instanceof CAhkDirectives || ah instanceof CAhkInclude
                 ? ah
                 : findDirectivesWithPos(ah.children, position);
         }
@@ -21,7 +22,7 @@ export function HoverDirectives(
     position: vscode.Position,
     AstRoot: TAstRoot,
 ): vscode.MarkdownString | undefined {
-    const ah: CAhkDirectives | undefined = findDirectivesWithPos(AstRoot, position);
+    const ah: CAhkDirectives | CAhkInclude | undefined = findDirectivesWithPos(AstRoot, position);
 
     return ah !== undefined
         ? DirectivesMDMap.get(ah.hashtag)