+++ src/provider/SymbolProvider/SymbolProvider.ts
@@ -1,5 +1,4 @@
 import type * as vscode from 'vscode';
-import type { CAhkFunc } from '../../AhkSymbol/CAhkFunc';
 import { useSymbolProvider } from '../../configUI';
 import { pm } from '../../core/ProjectManager';
 import { digDAFile } from '../../tools/DeepAnalysis/Diag/digDAFile';
@@ -8,8 +7,7 @@ import { getDAList } from '../../tools/DeepAnalysis/getDAList';
 function SymbolProviderCore(document: vscode.TextDocument): vscode.DocumentSymbol[] {
     const { AST, DocStrMap } = pm.updateDocDef(document);
 
-    const DAList: CAhkFunc[] = getDAList(AST);
-    digDAFile(DAList, document.uri, DocStrMap);
+    digDAFile(getDAList(AST), document.uri, DocStrMap);
 
     return useSymbolProvider()
         ? [...AST]